01-add-ledger
Creating Ledger, Account, Transaction, and Hints data structures
and creating unit tests for them.

A transaction has a date, src and dest accounts, and an amount.

An account has a type and a list of transactions from which it may
be a source or destination. Both accounts and transactions can be tagged.

Hints is a structure to read hints from an ini file. If a transaction
contains an account that isn't recognized, then hints may suggest a
name for it.

Ledger organizes instances of these structures. It maintains a
dictionary of accounts, and a list of transactions. These structures
contain references to each other. The ledger's job is to import
transactions from csv files and create the necessary accounts on
the fly.

This commit also contains unit tests.