master - refactor addTransaction
Refactored addTransaction to insert a copy of the provided transaction,
and return a reference to the ledger's internal reference. This
reference has its own src and dest references updated to the ledger's
internal accounts. This makes it trivial to add filtered transactions
from another ledger.

Also added an 'addTransactions' method to Ledger, which loadCsv now
uses. load, addTransactions, and loadCsv all return lists containing
references to the newly added transactions.