10-add-rpc
Created rpc server that can accept 3 functions; clear, dump, and load.
This server's main routine is in server.py.

daybook is no longer an interactive program; rather queries are sent
to the running server. parser.py was updated to reflect this. main.py
has been renamed to client.py to reflect this.

The programs are installed as daybook and daybookd.

Ledger: Changing Ledger.load method to accept a stream or string
of lines. Added loadCsvs method to handle loading a list of csvs.

Added configuration file which client.py and server.py read
from for the expected username, password, and listening port for the
server. The client will create this configuration file if it does
not exist on a first time run.

config.py contains functions to help set up daybook for a
first time run as well as reconcile command-line arguments against
confguration file parameters.