15-custom-hints-format
Hints file is no longer expected to be in ini format. Added a custom
parsing function to support colons in the variable names. By
coincidence, any character that isn't '=' is supported.

Line comments are also supported.

Changed the Hints constructor to only accept a hints argument, which may
be None, empty string, or a path to a hints file.

Changed usages of hintsini to simply be hints.