master
Empty src, test, or target fields will use 'void' for their account.
src, dest, and target fields can coexist.

src => (dest := target)
dest => (src := target)

And if target is the only heading present, then the sign of the first
number encountered in the 'amount' field determines if target is the
src or dest. This change allows amounts on a target to dictate their
currencies.

If none of (src, dest, target) are present, then 'this' is used for src
and dest.

Also adding unit tests.