16-types-in-names
Removing tags from Account.

Changed parsing logic with accounts. Accounts can now be created in a
transaction line like...

    old way: gasoline expense
    new way: expense.gasoline

And this will create an account name for gasoline of type expense.
This allows the Hints file to specify account types in the hints
themselves.

The presence of any spaces indicate something that should be fed into
Hints for a suggestion.

If an account of the same name (but possibly different type) is already
in the ledger, then the existing entry will be used. If the existing
account is a void account, then its type will be overwritten.

Updating unit tests and test data.