19-create-add-command
Created the "add" subcommand for appending a transaction to a local csv.

Add will load a hints file and a csv into a local ledger and use the
information to help autofill suggestions. Account names, tags, and
currencies can all be suggested based on the single csv and hints file.

Other changes:
  - Moved colonconf logic from Hints into its own file.
  - Added an optional arg to Ledger.load to allow it to skip
    invalid lines.
  - Ledger and Transaction now strips tags on entry