20-remove-target
Removed 'target' field from Ledger and updated Amount.createFromStr
to compensate. Amount no longer performs abs on amount numbers.

3 fields in amount are no longer valid. 1, 2, or 4 fields must be
specified.

- If 1 field was specified, then it should only be a numerical amount.
  The currency suggestion will be used.
- If 2 fields are specified, then one must be an amount and one must
  be a currency.
- If 4 fields are specified, then 2 amounts and 2 currencies must be
  specified.

In all cases, the first specified amount / currency always pertain
to src. If the amount is negative, then src is losing money and dest
is gaining.

Modifications to suggestAccount to fix a bug unrelated to this feature
with how thisname was processed.

Updates to unit-tests and resources as necessary to test the new
behaviors.