21-enforce-account-name-presence
If a src or dest field is in the csv, then lines are required to
specify account names. No line can only specify an account type or be
empty.

This also required a bit of a fix to Account.createFromStr, as it was
possible to accidentally use an account like...

    'void.asset'

Which makes no sense. Minor bug fix there, and the parsing should be a
bit more optimized.

An exception is made for an entry 'void'. That will use an account
'void.void'

Fixed unit tests in test_ledger as well.