26-remove-last-currency
Removing "last currency" behavior. Previously, the ledger would keep
track of which currencies were used last, and would suggest their use
in the event a currency wasn't specified in the transaction.

Because of a massive headache this accidentally caused for me while
book-keeping, this "helpful" feature is now removed. It is also
uncommon and unintuitive for a CSV to suddenly stop mentioning
currencies. It actually increases manual work to do this.

The ledger's primary currency will now be the only suggestion.