28-auto-orient-transactions
Transaction src and dest info is oriented automatically. Under the hood,
the Amount.src_amount field must now always be negative. The Ledger.load
method will check the src_amount and if the src_amount is positive, then
the Amount will be corrected and the src and dest accounts will be
swapped before being committed to the ledger.

This makes intuitive sense that money should flow from src to dest, and
it still allows users to use positive amounts in the "dest" fields of
CSVs to indicate income to the src account.

It also allows for easier duplicate transaction detection, as src and
dest accounts, amounts, and currencies will automatically be oriented
upon entry.