31-refactor config
Refactored how config args are read and where to add new config args.
This makes it more intuitive to add new configuration file entries.

To add a new config parameter, make it an entry in ./etc/default.ini and
update the get_defaults() function in config.py as well as the code in
do_first_time_setup if special logic is necessary.