30-detect-duplicates
Implemented automatic duplicate detection. Transactions will be flagged
as duplicates if they are loaded from a unique "perspective" and  whose
date is within a specificed date range of another group of transactions
that only has one date. This behavior is configurable via daybook.ini
and the --duplicate-window option.

Transactions can be loaded from a "perspective". This is very likely
the basename of the CSV. The gist of this automatic duplicate detection
is to group transactions that are identical except for their perspective
and their date. The idea is that if two nearly identical transactions
are loaded from different perspectives with a close enough date, that
those two transactions are likely the same event. eg. asset.checking
sends 500:usd to asset.investment, but asset.investment doesn't see the
money until a few days later. This event should not be reported twice.

If two similar transactions are loaded from the same perspective, then
these are never treated as duplicates. For example, if asset.checking is
used to purhcase the same item from the same vendor twice in one day, or
within a few days, these 2 events will very likely be recorded from CSVs
each called "asset.checking.csv". Because the "perspective" is the same,
these events are treated as unique transactions.