DAYBOOK_7: Removing system default ini from install
The system wide default ini is no longer installed. It is instead stored
as a docstring within the program.

Python may install programs to a variety of locations. Since the default
ini file was potentially referenced in flight I'd rather not keep track
of its location on disk.

The ini is also created with comments in it.