Added report subcommand
This subcommand is similar to the convert sucommand. It points at a
python3 file and expects it to have a certain function.

This "report" function accepts a ledger and budget reference and is
expected to return a string that the user will format based on arbitrary
logic involving the budget and ledger references.

This also refactors the importer and load_budget functions into common
code, and adds the necessary manpages and unittests.