DAYBOOK_10: Report and convert modules as subcommands
Removed the --list and --describe option from convert and report in
favor of dynamically finding convert and report modules and treating
each found one as a subcommand.

This means each gets a --help argument, so the help and description
messages are displayed automatically.

All of the options for loading CSVS are now given to the report modules
themseves, so the daybook report help message is minimal and shows just
the available report modules.