DAYBOK_14: Simpler filtering
All filter options have been replaced by a single --filter option which
accepts a string to be evaluated as a python3 conditional.

The Ledger.getTransactions method can now accept both that string or a
function reference that accepts a transaction.

Existing date parsing has been removed in favor of the superdate module
parsing. Under the hood transaction.date is stored as a SuperDate, but
that should still preserve the standard python3 datetime interface.