feature: Added docker container
A docker container was added because some operating environments cannot
install daybook using the `pip install --user .` method.

This is because the convert and report modules must be installed in a
user directory.

A proper fix would probably be to import the modules more robustly or
print a more helpful error message, or perhaps design an integration
of the presets into daybook properly.

But this would be annoying all the same as the man pages and completion
script also require a user directory to install, which wouldn't be
captured by the venv anyway.

So a docker container has been created to provide a compatible
interactive environment; if a bit heavy handed as an unminized Ubuntu
container is ~1Gb in size.